Access the regitry through a reverse proxy apache virtual host.
To access the registry host from outside the VPC we install and setup an
apache virtual host in the bastion host.

The registry is accessed using the DNS name registry.taletoul.com which
is created by terraform during infrastructure creation.

The reverse proxy is used to access the web interface and the docker
registry with the docker client.

Signed-off-by: Tale Toul <tale.toul@gmail.com>